refactor: add abstract class for views.
