+++ src/view/event-edit-view.js
@@ -1,4 +1,4 @@
-import { createElement } from '../render.js';
+import AbstractView from '../framework/view/abstract-view.js';
 import { DateFormat, EVENT_TYPES } from '../const.js';
 import { getFormattedDate, getCapitalizedString, getHtmlId } from '../utils.js';
 
@@ -138,7 +138,7 @@ function createEventEditTemplate(
   );
 }
 
-export default class EventEditView {
+export default class EventEditView extends AbstractView {
   viewId = null;
   event = null;
   currentDestination = null;
@@ -153,6 +153,7 @@ export default class EventEditView {
     currentOffersPack,
     allDestinations,
     allOffersPacks}){
+    super();
     this.viewId = viewId;
     this.event = event;
     this.currentDestination = currentDestination;
@@ -161,7 +162,7 @@ export default class EventEditView {
     this.allOffersPacks = allOffersPacks;
   }
 
-  getTemplate() {
+  get template() {
     return createEventEditTemplate(
       this.viewId,
       this.event,
@@ -170,16 +171,5 @@ export default class EventEditView {
       this.allDestinations,
       this.allOffersPacks,
     );
-  }
-
-  getElement() {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
-    }
-    return this.element;
-  }
-
-  removeElement() {
-    this.element = null;
   }
 }
