+++ src/view/trip-view.js
@@ -1,4 +1,4 @@
-import { createElement } from '../render.js';
+import AbstractView from '../framework/view/abstract-view.js';
 
 function createTripTemplate() {
   return (
@@ -8,19 +8,8 @@ function createTripTemplate() {
   );
 }
 
-export default class TripView {
-  getTemplate() {
+export default class TripView extends AbstractView {
+  get template() {
     return createTripTemplate();
-  }
-
-  getElement() {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
-    }
-    return this.element;
-  }
-
-  removeElement() {
-    this.element = null;
   }
 }
